feat: add pages to atomic structure
- add details page & section
- add settings page & section
- add create-task page & section
- change Header|Footer to TheHeader & TheFooter
- created initial structure of atomic design - app router
