+++ src/app/details/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+import { DetailsTaskSection } from "@/components/04-templates";
+
+export const DetailsTask = () => {
+  return (
+    <div>
+      <DetailsTaskSection />
+    </div>
+  );
+};
+
+export default DetailsTask;
