+++ src/app/settings/create-task/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+import { CreateTaskSection } from "@/components/04-templates";
+
+export const Create = () => {
+  return (
+    <div>
+      <CreateTaskSection data={undefined} />
+    </div>
+  );
+};
+
+export default Create;
