+++ src/app/settings/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+import { AdminOptionsSection } from "@/components/04-templates";
+
+export const Settings = () => {
+  return (
+    <div>
+      <AdminOptionsSection />
+    </div>
+  );
+};
+
+export default Settings;
