+++ src/components/02-molecules/TheHeader.tsx
@@ -7,11 +7,11 @@ import {
   styled,
   useTheme,
 } from "@mui/material";
-import Logo from "@/components/01-atoms/LogoSign";
 import HeaderUserbox from "@/layouts/BaseLayout/Header/Userbox";
 import HeaderSearch from "@/layouts/BaseLayout/Header/Search";
 import HeaderUserConnect from "@/layouts/BaseLayout/Header/UserConnect";
 import { useConnect, useDisconnect } from "wagmi";
+import { Logo } from "../01-atoms/LogoSign";
 
 const HeaderWrapper = styled(Box)(
   ({ theme }) => `
@@ -28,10 +28,10 @@ const HeaderWrapper = styled(Box)(
             left: ${theme.sidebar.width};
             width: auto;
         }
-`
+`,
 );
 
-function Header({ data }) {
+export function TheHeader({ data }) {
   const theme = useTheme();
   const logoImage = "/static/images/logo/logo-" + theme.palette.mode + ".svg";
 
@@ -53,14 +53,14 @@ function Header({ data }) {
           theme.palette.mode === "dark"
             ? `0 1px 0 ${alpha(
                 lighten(theme.colors.primary.main, 0.7),
-                0.15
+                0.15,
               )}, 0px 2px 8px -3px rgba(0, 0, 0, 0.2), 0px 5px 22px -4px rgba(0, 0, 0, .1)`
             : `0px 2px 8px -3px ${alpha(
                 theme.colors.alpha.black[100],
-                0.2
+                0.2,
               )}, 0px 5px 22px -4px ${alpha(
                 theme.colors.alpha.black[100],
-                0.1
+                0.1,
               )}`,
       }}
     >
@@ -99,5 +99,3 @@ function Header({ data }) {
     </HeaderWrapper>
   );
 }
-
-export default Header;
