+++ src/components/02-molecules/index.tsx
@@ -2,5 +2,5 @@ export * from "./CoverAdminOptions";
 export * from "./CoverCreateTask";
 export * from "./CoverHomeTasks";
 export * from "./CoverUserProfile";
-export * from "./Footer";
-export * from "./Header";
+export * from "./TheFooter";
+export * from "./TheHeader";
