+++ src/components/04-templates/index.tsx
@@ -1,4 +1,7 @@
 export * from "./404Section";
 export * from "./500Section";
-export * from "./ProfileSection";
+export * from "./AdminOptionsSection";
+export * from "./CreateTaskSection";
+export * from "./DetailsTaskSection";
 export * from "./HomeSection";
+export * from "./ProfileSection";
