+++ src/layouts/SidebarLayout/index.tsx
@@ -4,7 +4,7 @@ import { Outlet } from "react-router-dom";
 
 import Sidebar from "./Sidebar";
 import Header from "./Header";
-import Footer from "@/components/02-molecules/Footer";
+import Footer from "@/components/02-molecules/TheFooter";
 interface SidebarLayoutProps {
   children?: ReactNode;
 }
@@ -29,14 +29,14 @@ const SidebarLayout: FC<SidebarLayoutProps> = () => {
               theme.palette.mode === "dark"
                 ? `0 1px 0 ${alpha(
                     lighten(theme.colors.primary.main, 0.7),
-                    0.15
+                    0.15,
                   )}, 0px 2px 4px -3px rgba(0, 0, 0, 0.2), 0px 5px 12px -4px rgba(0, 0, 0, .1)`
                 : `0px 2px 4px -3px ${alpha(
                     theme.colors.alpha.black[100],
-                    0.1
+                    0.1,
                   )}, 0px 5px 12px -4px ${alpha(
                     theme.colors.alpha.black[100],
-                    0.05
+                    0.05,
                   )}`,
           },
         }}
